requirements.txt
APScheduler==3.10.4
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
cssselect2==0.7.0
diff-match-patch==20241021
dj-database-url==3.0.0
Django==4.2.11
django-apscheduler==0.7.0
django-auditlog==3.0.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
django-haystack==3.3.0
django-import-export==4.2.1
django-jsonfield==1.4.1
django-mathfilters==1.0.0
django-model-utils==5.0.0
django-simple-history==3.7.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
et_xmlfile==2.0.0
future==1.0.0
gunicorn==23.0.0
html5lib==1.1
idna==3.10
inflection==0.5.1
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==3.0.2
numpy==2.1.3
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pyHanko==0.25.2
pyhanko-certvalidator==0.26.4
PyJWT==2.9.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pypdf==5.1.0
pypng==0.20220715.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzk==0.9
qrcode==8.0
reportlab==4.0.9
requests==2.32.3
responses==0.25.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
svglib==1.5.1
swapper==1.4.0
tablib==3.7.0
tinycss2==1.4.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.3
utils==1.0.2
webencodings==0.5.1
whitenoise==6.8.2
Whoosh==2.7.4
xhtml2pdf==0.2.16
XlsxWriter==3.2.0
